HW_Report/ex/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
google-generativeai
python-socketio
eventlet
django-filter
drf-yasg